services/auth-service/requirements.txt
flask
pyodbc
python-dotenv
bcrypt
pyjwt